+++ email/tests/smtp.rs
@@ -39,8 +39,8 @@ async fn test_smtp_features() {
         ..Default::default()
     };
 
-    let imap_ctx = ImapContextBuilder::new(account_config.clone(), imap_config);
-    let smtp_ctx = SmtpContextBuilder::new(account_config.clone(), smtp_config);
+    let imap_ctx = ImapContextBuilder::new(imap_config);
+    let smtp_ctx = SmtpContextBuilder::new(smtp_config);
     let backend_builder = BackendBuilder::new(account_config.clone(), (imap_ctx, smtp_ctx))
         .with_purge_folder(|ctx| PurgeImapFolder::some_new_boxed(&ctx.0))
         .with_list_envelopes(|ctx| ListImapEnvelopes::some_new_boxed(&ctx.0))